refactor: clarify plus adapter structure
