+++ src/components/plus/device/DeviceControlTools.vue
